+++ Sources/Runner/main.swift
@@ -9,10 +9,9 @@ do {
     let isVerbose = CommandLine.arguments.contains("--verbose") || (ProcessInfo.processInfo.environment["DEBUG"] != nil)
     let isSilent = CommandLine.arguments.contains("--silent")
     let logger = Logger(isVerbose: isVerbose, isSilent: isSilent)
+    logger.debug("Launching Danger Swift (v\(DangerVersion))")
 
     if cliLength > 1 {
-        logger.debug("Launching Danger Swift \(CommandLine.arguments[1]) (v\(DangerVersion))")
-
         switch(CommandLine.arguments[1]) {
         case "ci", "local", "pr":
             let exitCode = try runDangerJSCommandToRunDangerSwift(CommandLine.arguments[1], logger: logger)
@@ -26,11 +25,14 @@ do {
             fatalError("Danger Swift does not support this argument, it only handles ci, local, pr & edit'")
         }
     } else {
-        logger.logInfo("Deprecated: Please don't use 'danger-swift' on its own to evaluate a Dangerfile.swift")
-        logger.logInfo("            you can change it to `danger-swft runner` - this will get removed when")
-        logger.logInfo("            danger hits 1.0 and replaced by help info")
+        // [SOON]
+        // but I can't think of a way to get Danger JS to handle the sub-commands from Danger Swift yet
+        //
+        // logger.logInfo("Deprecated: Please don't use 'danger-swift' on its own to evaluate a Dangerfile.swift")
+        // logger.logInfo("            you can change it to `danger-swft runner` - this will get removed when")
+        // logger.logInfo("            danger hits 1.0 and replaced by help info")
+        //
         try getDSLData(logger: logger, runDanger)
-
     }
 } catch {
     exit(1)